+++ Code/LYRUIConversationListViewController.m
@@ -8,24 +8,19 @@
 
 #import <objc/runtime.h>
 #import "LYRUIConversationListViewController.h"
-#import "LYRUIDataSourceChange.h"
-#import "LYRUIConstants.h"
 
-@interface LYRUIConversationListViewController () <UISearchBarDelegate, UISearchDisplayDelegate, LYRQueryControllerDelegate>
+static NSString *const LYRUIConversationCellReuseIdentifier = @"LYRUIConversationCellReuseIdentifier";
 
-@property (nonatomic) UISearchBar *searchBar;
-@property (nonatomic) UISearchDisplayController *searchController;
-@property (nonatomic) NSMutableArray *filteredConversations;
-@property (nonatomic) NSPredicate *searchPredicate;
+@interface LYRUIConversationListViewController () <UIActionSheetDelegate, LYRQueryControllerDelegate>
+
+@property (nonatomic) LYRClient *layerClient;
 @property (nonatomic) LYRQueryController *queryController;
-@property (nonatomic) NSMutableArray *objectChages;
-@property (nonatomic) BOOL isOnScreen;
+@property (nonatomic) BOOL hasAppeared;
+@property (nonatomic) LYRConversation *conversationToDelete;
 
 @end
 
 @implementation LYRUIConversationListViewController
-
-static NSString *const LYRUIConversationCellReuseIdentifier = @"conversationCellReuseIdentifier";
 
 + (instancetype)conversationListViewControllerWithLayerClient:(LYRClient *)layerClient
 {
@@ -37,10 +32,7 @@ static NSString *const LYRUIConversationCellReuseIdentifier = @"conversationCell
 {
     self = [super initWithStyle:UITableViewStylePlain];
     if (self)  {
-        // Set property from designated initializer
         _layerClient = layerClient;
-        
-        // Set default configuration for public configuration properties
         _cellClass = [LYRUIConversationTableViewCell class];
         _displaysConversationImage = YES;
         _allowsEditing = YES;
@@ -49,84 +41,57 @@ static NSString *const LYRUIConversationCellReuseIdentifier = @"conversationCell
     return self;
 }
 
-- (id) init
+- (id)init
 {
     [NSException raise:NSInternalInconsistencyException format:@"Failed to call designated initializer"];
     return nil;
 }
 
-#pragma mark - VC Lifecycle Methods
+#pragma mark - Lifecycle
+
 - (void)viewDidLoad
 {
     [super viewDidLoad];
-    // Accessibility
+
     self.title = @"Messages";
     self.accessibilityLabel = @"Messages";
     self.tableView.accessibilityLabel = @"Conversation List";
-    
-    // Search Bar Setup
-//    self.searchBar = [[UISearchBar alloc] initWithFrame:CGRectZero];
-//    self.searchBar.accessibilityLabel = @"Search Bar";
-//    self.searchBar.delegate = self;
-//    self.searchController = [[UISearchDisplayController alloc] initWithSearchBar:self.searchBar contentsController:self];
-//    self.searchController.delegate = self;
-//    self.searchController.searchResultsDelegate = self;
-//    self.searchController.searchResultsDataSource = self;
 
-    // Set Search Bar as Table View Header
-    //self.tableView.tableHeaderView = self.searchBar;
-    //[self.tableView setContentOffset:CGPointMake(0, 44)];
-
-    // DataSoure
     [self setupConversationDataSource];
 }
 
 - (void)viewWillAppear:(BOOL)animated
 {
     [super viewWillAppear:animated];
-    self.isOnScreen = TRUE;
-    
-    // Set public configuration properties once view has loaded
-    [self.tableView registerClass:self.cellClass forCellReuseIdentifier:LYRUIConversationCellReuseIdentifier];
-    [self.searchController.searchResultsTableView registerClass:self.cellClass forCellReuseIdentifier:LYRUIConversationCellReuseIdentifier];
-    
-    self.tableView.rowHeight = self.rowHeight;
-    self.searchController.searchResultsTableView.rowHeight = self.rowHeight;
-    
-    if (self.allowsEditing) {
-        [self addEditButton];
+
+    if (!self.hasAppeared) {
+        // Set public configuration properties once view has loaded
+        [self.tableView registerClass:self.cellClass forCellReuseIdentifier:LYRUIConversationCellReuseIdentifier];
+        self.tableView.rowHeight = self.rowHeight;
+        if (self.allowsEditing) {
+            [self addEditButton];
+        }
+        self.hasAppeared = YES;
     }
-    [self.tableView reloadData];
 }
 
-- (void)viewDidDisappear:(BOOL)animated
-{
-    [super viewDidDisappear:animated];
-    self.isOnScreen = NO;
-}
-
-#pragma mark - Public setters
+#pragma mark - Public Setters
 
 - (void)setAllowsEditing:(BOOL)allowsEditing
 {
-    if (self.isOnScreen) {
-        @throw [NSException exceptionWithName:NSInternalInconsistencyException reason:@"Cannot set editing mode after the view has been loaded" userInfo:nil];
+    if (self.hasAppeared) {
+        @throw [NSException exceptionWithName:NSInternalInconsistencyException reason:@"Cannot change editing mode after the view has been presented" userInfo:nil];
     }
     _allowsEditing = allowsEditing;
-
-    if (self.navigationItem.leftBarButtonItem && !allowsEditing) {
-        self.navigationItem.leftBarButtonItem = nil;
-    }
 }
 
 - (void)setCellClass:(Class<LYRUIConversationPresenting>)cellClass
 {
-    if (self.isOnScreen) {
-        @throw [NSException exceptionWithName:NSInternalInconsistencyException reason:@"Cannot set cellClass after the view has been loaded" userInfo:nil];
+    if (self.hasAppeared) {
+        @throw [NSException exceptionWithName:NSInternalInconsistencyException reason:@"Cannot change cell class after the view has been presented" userInfo:nil];
     }
-    
     if (!class_conformsToProtocol(cellClass, @protocol(LYRUIConversationPresenting))) {
-        @throw [NSException exceptionWithName:NSInternalInconsistencyException reason:@"Cell class cellClass must conform to LYRUIConversationPresenting Protocol" userInfo:nil];
+        @throw [NSException exceptionWithName:NSInternalInconsistencyException reason:@"Cell class must conform to LYRUIConversationPresenting" userInfo:nil];
 
     }
     _cellClass = cellClass;
@@ -134,32 +99,28 @@ static NSString *const LYRUIConversationCellReuseIdentifier = @"conversationCell
 
 - (void)setRowHeight:(CGFloat)rowHeight
 {
-    if (self.isOnScreen) {
-        @throw [NSException exceptionWithName:NSInternalInconsistencyException reason:@"Cannot set rowHeight after the view has been loaded" userInfo:nil];
+    if (self.hasAppeared) {
+        @throw [NSException exceptionWithName:NSInternalInconsistencyException reason:@"Cannot change row height after the view has been presented" userInfo:nil];
     }
     _rowHeight = rowHeight;
 }
 
 - (void)setDisplaysConversationImage:(BOOL)displaysConversationImage
 {
-    if (self.isOnScreen) {
-        @throw [NSException exceptionWithName:NSInternalInconsistencyException reason:@"Cannot set displaysConversationImage after the view has been loaded" userInfo:nil];
+    if (self.hasAppeared) {
+        @throw [NSException exceptionWithName:NSInternalInconsistencyException reason:@"Cannot change conversation image display after the view has been presented" userInfo:nil];
     }
     _displaysConversationImage = displaysConversationImage;
 }
 
-#pragma mark - Navigation Bar Edit Button
+#pragma mark - Set Up
 
 - (void)addEditButton
 {
-    UIBarButtonItem *editButtonItem = [[UIBarButtonItem alloc] initWithTitle:@"Edit"
-                                                                       style:UIBarButtonItemStylePlain
-                                                                      target:self
-                                                                      action:@selector(editButtonTapped)];
-    editButtonItem.accessibilityLabel = @"Edit Button";
-    self.navigationItem.leftBarButtonItem = editButtonItem;
+    if (self.navigationItem.leftBarButtonItem) return;
+    self.editButtonItem.accessibilityLabel = @"Edit Button";
+    self.navigationItem.leftBarButtonItem = self.editButtonItem;
 }
-
 
 - (void)setupConversationDataSource
 {
@@ -168,48 +129,12 @@ static NSString *const LYRUIConversationCellReuseIdentifier = @"conversationCell
     
     self.queryController = [self.layerClient queryControllerWithQuery:query];
     self.queryController.delegate = self;
-    NSError *error = nil;
+    NSError *error;
     BOOL success = [self.queryController execute:&error];
     if (!success) NSLog(@"LayerKit failed to execute query with error: %@", error);
-    [self.tableView reloadData];
 }
 
-- (void)reloadConversations
-{
-    if (self.searchController.active) {
-        [self.searchController.searchResultsTableView reloadData];
-    } else {
-        [self.tableView reloadData];
-    }
-}
-
-- (BOOL)isSearching
-{
-    return self.searchController.active;
-}
-
-#pragma mark - UISearchDisplayDelegate Methods
-
-- (void)searchDisplayControllerWillBeginSearch:(UISearchDisplayController *)controller
-{
-    // We react to search begining
-}
-
-- (void)searchDisplayControllerDidEndSearch:(UISearchDisplayController *)controller
-{
-    // We respond to ending the search
-}
-
-- (void)searchBar:(UISearchBar *)searchBar textDidChange:(NSString *)searchText
-{
-    if ([self.dataSource respondsToSelector:@selector(conversationListViewController:didSearchWithString:completion:)]) {
-        [self.dataSource conversationListViewController:self didSearchWithString:searchText completion:^(NSSet *conversations) {
-            [self.tableView reloadData];
-        }];
-    }
-}
-
-#pragma mark - Table view data source methods
+#pragma mark - UITableViewDataSource
 
 - (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
 {
@@ -223,18 +148,18 @@ static NSString *const LYRUIConversationCellReuseIdentifier = @"conversationCell
     return conversationCell;
 }
 
-/**
- 
- LAYER - Extracting content from an LYRConversation object so that it can be displayed in a tableViewCell.
- 
- */
+- (BOOL)tableView:(UITableView *)tableView canEditRowAtIndexPath:(NSIndexPath *)indexPath
+{
+    return self.allowsEditing;
+}
+
+#pragma mark - Cell Configuration
+
 - (void)configureCell:(UITableViewCell<LYRUIConversationPresenting> *)conversationCell atIndexPath:(NSIndexPath *)indexPath
 {
-    // Present Conversation
     LYRConversation *conversation = [self.queryController objectAtIndexPath:indexPath];
     [conversationCell presentConversation:conversation];
     
-    // Update cell with image if needed
     if (self.displaysConversationImage) {
         if ([self.dataSource respondsToSelector:@selector(conversationListViewController:imageForConversation:)]) {
             UIImage *conversationImage = [self.dataSource conversationListViewController:self imageForConversation:conversation];
@@ -244,11 +169,9 @@ static NSString *const LYRUIConversationCellReuseIdentifier = @"conversationCell
         }
     }
     
-    // Update Cell with unread message count
     LYRRecipientStatus status = [[conversation.lastMessage.recipientStatusByUserID objectForKey:self.layerClient.authenticatedUserID] integerValue];
     [conversationCell updateWithLastMessageRecipientStatus:status];
     
-    // Update Cell with Label
     if ([self.dataSource respondsToSelector:@selector(conversationListViewController:labelForConversation:)]) {
         NSString *conversationLabel = [self.dataSource conversationListViewController:self labelForConversation:conversation];
         [conversationCell updateWithConversationLabel:conversationLabel];
@@ -257,10 +180,7 @@ static NSString *const LYRUIConversationCellReuseIdentifier = @"conversationCell
     }
 }
 
-- (BOOL)tableView:(UITableView *)tableView canEditRowAtIndexPath:(NSIndexPath *)indexPath
-{
-    return YES;
-}
+#pragma mark - UITableViewDelegate
 
 - (NSArray *)tableView:(UITableView *)tableView editActionsForRowAtIndexPath:(NSIndexPath *)indexPath {
     
@@ -277,79 +197,41 @@ static NSString *const LYRUIConversationCellReuseIdentifier = @"conversationCell
     return @[globalDeleteAction, localDeleteAction];
 }
 
-/**
- 
- LAYER - Deleting a Layer Conversation
- 
- */
-- (void)deleteConversationAtIndexPath:(NSIndexPath *)indexPath withDeletionMode:(LYRDeletionMode)deletionMode
-{
-    LYRConversation *conversation = [self.queryController objectAtIndexPath:indexPath];
-    NSError *error;
-    BOOL success = [conversation delete:deletionMode error:&error];
-    if (!success) {
-        if ([self.delegate respondsToSelector:@selector(conversationListViewController:didFailDeletingConversation:deletionMode:error:)]) {
-            [self.delegate conversationListViewController:self didFailDeletingConversation:conversation deletionMode:deletionMode error:error];
-        }
-    } else {
-        if ([self.delegate respondsToSelector:@selector(conversationListViewController:didDeleteConversation:deletionMode:)]) {
-            [self.delegate conversationListViewController:self didDeleteConversation:conversation deletionMode:deletionMode];
-        }
-    }
-}
-
 - (void)tableView:(UITableView *)tableView commitEditingStyle:(UITableViewCellEditingStyle)editingStyle forRowAtIndexPath:(NSIndexPath *)indexPath
 {
-    // Not implemented
+    self.conversationToDelete = [self.queryController objectAtIndexPath:indexPath];
+    UIActionSheet *actionSheet = [[UIActionSheet alloc] initWithTitle:nil delegate:self cancelButtonTitle:@"Cancel" destructiveButtonTitle:@"Global" otherButtonTitles:@"Local", nil];
+    [actionSheet showInView:self.view];
 }
-
-#pragma mark - Table view delegate methods
 
 - (void)tableView:(UITableView *)tableView didSelectRowAtIndexPath:(NSIndexPath *)indexPath
 {
-    LYRConversation *conversation = [self.queryController objectAtIndexPath:indexPath];
     if ([self.delegate respondsToSelector:@selector(conversationListViewController:didSelectConversation:)]){
+        LYRConversation *conversation = [self.queryController objectAtIndexPath:indexPath];
         [self.delegate conversationListViewController:self didSelectConversation:conversation];
     }
 }
 
-- (CGFloat)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
+#pragma mark - UIActionSheetDelegate
+
+- (void)actionSheet:(UIActionSheet *)actionSheet clickedButtonAtIndex:(NSInteger)buttonIndex
 {
-    return self.rowHeight;
+    if (buttonIndex == actionSheet.destructiveButtonIndex) {
+        [self deleteConversation:self.conversationToDelete withDeletionMode:LYRDeletionModeAllParticipants];
+    } else if (buttonIndex == actionSheet.firstOtherButtonIndex) {
+        [self deleteConversation:self.conversationToDelete withDeletionMode:LYRDeletionModeLocal];
+    } else if (buttonIndex == actionSheet.cancelButtonIndex) {
+        [self setEditing:NO animated:YES];
+    }
+    self.conversationToDelete = nil;
 }
 
-#pragma mark - Conversation Editing Methods
-
-// Set table view into editing mode and change left bar buttong to a done button
-- (void)editButtonTapped
-{
-    [self.tableView setEditing:TRUE animated:TRUE];
-    UIBarButtonItem *doneButtonItem = [[UIBarButtonItem alloc] initWithTitle:@"Done"
-                                                                       style:UIBarButtonItemStylePlain
-                                                                      target:self
-                                                                      action:@selector(doneButtonTapped)];
-    doneButtonItem.accessibilityLabel = @"Done";
-    self.navigationItem.leftBarButtonItem = doneButtonItem;
-}
-
-- (void)doneButtonTapped
-{
-    [self.tableView setEditing:NO animated:YES];
-    [self addEditButton];
-}
-
-- (UITableViewCellEditingStyle)tableView:(UITableView *)tableView editingStyleForRowAtIndexPath:(NSIndexPath *)indexPath
-{
-    return UITableViewCellEditingStyleDelete;
-}
-
-#pragma mark - Conversation Query Controller 
+#pragma mark - LYRQueryControllerDelegate
 
 - (void)queryControllerWillChangeContent:(LYRQueryController *)queryController
 {
     [self.tableView beginUpdates];
 }
-
 
 - (void)queryController:(LYRQueryController *)controller
         didChangeObject:(id)object
@@ -384,6 +266,29 @@ static NSString *const LYRUIConversationCellReuseIdentifier = @"conversationCell
 - (void)queryControllerDidChangeContent:(LYRQueryController *)queryController
 {
     [self.tableView endUpdates];
+}
+
+#pragma mark - Helpers
+
+- (void)deleteConversationAtIndexPath:(NSIndexPath *)indexPath withDeletionMode:(LYRDeletionMode)deletionMode
+{
+    LYRConversation *conversation = [self.queryController objectAtIndexPath:indexPath];
+    [self deleteConversation:conversation withDeletionMode:deletionMode];
+}
+
+- (void)deleteConversation:(LYRConversation *)conversation withDeletionMode:(LYRDeletionMode)deletionMode
+{
+    NSError *error;
+    BOOL success = [conversation delete:deletionMode error:&error];
+    if (!success) {
+        if ([self.delegate respondsToSelector:@selector(conversationListViewController:didFailDeletingConversation:deletionMode:error:)]) {
+            [self.delegate conversationListViewController:self didFailDeletingConversation:conversation deletionMode:deletionMode error:error];
+        }
+    } else {
+        if ([self.delegate respondsToSelector:@selector(conversationListViewController:didDeleteConversation:deletionMode:)]) {
+            [self.delegate conversationListViewController:self didDeleteConversation:conversation deletionMode:deletionMode];
+        }
+    }
 }
 
 @end