+++ CHANGELOG.md
@@ -40,6 +40,8 @@
 * Changed `LYRUIPaticipantSectionHeaderView` to `LYRUIParticipantSectionHeaderView`.
 * Removed `initWithKey:` from `LYRUIParticipantSectionHeaderView`.
 * Changed `keyLabel` property to `nameLabel` on `LYRUIParticipantSectionHeaderView`.
+* Removed `conversationListViewController:didSearchWithString:completion:` from `LYRUIConversationListViewControllerDataSource`.
+* Changed `layerClient` property on `LYRUIConversationListViewController` from `readwrite` to `readonly`.
 
 ### Bug Fixes